+++ CHANGELOG.md
@@ -39,6 +39,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Add a wizard component [#725](https://github.com/azavea/echo-locator/pull/725)
 - Add Your Trips section and commute maps [#724](https://github.com/azavea/echo-locator/pull/724)
 - Add transit trip directions overlay [#726](https://github.com/azavea/echo-locator/pull/726)
+- Add user profile wizard [#729](https://github.com/azavea/echo-locator/pull/729)
 
 ### Changed